my_function/requirements.txt
functions-framework==3.*
torch
openai-whisper
requests
ffmpeg-python
faster-whisper